WIP: Loop table rule S1 calculation + Tests

+++ src/main/kotlin/tin/model/v2/transducer/TransducerEdge.kt
@@ -8,7 +8,7 @@ class TransducerEdge(
     override var target: Node,
     var incomingString: String,
     var outgoingString: String,
-    var cost: Double
+    var cost: Int
 ) : Edge(
         source, target, "$incomingString, $outgoingString, $cost"
 ){
